+++ projects/glassui/src/lib/content/content.module.ts
@@ -1,17 +1,20 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {CardComponent} from './card/card.component';
+import { TableComponent } from './table/table/table.component';
 
 
 @NgModule({
   declarations: [
-    CardComponent
+    CardComponent,
+    TableComponent
   ],
   imports: [
     CommonModule
   ],
   exports: [
-    CardComponent
+    CardComponent,
+    TableComponent
   ]
 })
 export class ContentModule {