+++ projects/glassui/src/public-api.ts
@@ -18,6 +18,7 @@ export * from './lib/navigation/page/page.component';
 //Content
 export * from './lib/content/content.module';
 export * from './lib/content/card/card.component';
+export * from './lib/content/table/table/table.component';
 
 //toast
 export * from './lib/toast/toast.module';